pyproject.toml
[tool.poetry]
name = "chorus"
version = "0.1.0"
description = ""
authors = ["Toan Phung <toan.phung@proteo.info>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
django = "^4.2.5"
plotly = "^5.17.0"
pandas = "^2.1.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.7"
drf-url-filters = "^0.5.1"
drf-flex-fields = "^1.0.2"
drf-spectacular = "^0.26.4"
dj-rest-auth = "^5.0.1"
django-request = "^1.6.3"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
boto3 = "^1.28.52"
django-storages = "^1.14"
channels-redis = {extras = ["cryptography"], version = "^4.1.0"}
django-dbbackup = "^4.0.2"
django-cors-headers = "^4.2.0"
django-rq = "^2.8.1"
django-redis = "^5.3.0"
uniprotparser = "^1.1.0"


[tool.poetry.group.dev.dependencies]
xlrd = "^2.0.1"
openpyxl = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
